Bond Manager (#3)
* feat(bond): Add bond manager

When a Sequencer wants to appendBatch, they need to have previously
called BondManager.deposit in order to have stake available to be locked.

Inside Transitioners, whenever somebody provides a witness, their contribution
for that dispute (keyed by pre-state root) is stored. Later on, once the dispute
is resolved, the fraud verifier makes a call to the BondManager and if there was
no fraud, then the sequencer's bond is returned, otherwise it is distributed
proportionally to all the witness providers that helped resolve that dispute

* test(bond): add unit tests for bond manager

* fix(state-transitioner): reward `applyTransaction` callers

* chore(bond-manager): s/payout/claim and fix typo

* feat(bond-manager): add delayed withdrawals for publishers to allow fraud proving

* fix(bond-manager): improve error strings

* test(fraud-verifier): mock bondmanager

* test(state-comm-chain): mock bondmanager

* test(state-transitioner): mock bondmanager

* chore: s/forceInclusion/dispute

* chore: s/stake/isCollateralized

* test(bond-manager): smodify ovmFraudVerifier and remove TODO from contract

* fix: re-arranges contracts to match desired styling

also removes the require modifier from isCollateralized, allowing us to remove the state commitment chain state variable altogether

* fix: call finalize from a fraud verifier mock

Since smoddit does not allow us to mod storage set at the constructor, and we don't want
to have to call the finalizeFraudVerification function on the OVM_FraudVerifier, we will just
use our own Fraud Verifier mock that we use for mocking the state transitioners and simulating
the call to BondManager.finalize

* feat(fraud-verifier): reward callers of initialize/finalizeFraudVerification

NOTE: This requires that the address of the BondManager is already set in the AddressManager. However, since the BondManager also needs the address of the FraudVerifier, this creates a cyclic dependency. To solve it, you MUST pre-calculate the address of both contracts before deployment (by hashing the deployer address + nonce) and set that in the AddressManager

* test(bond-manager): ensure isCollateralized is false after withdrawing

* chore: remove unused parameter in isCollateralized

* chore: address review comments and add more docs

* feat(bond-manager): track user state instead of numbers

* feat(bond-manager): use gas-weighted rewards

Previously, we'd reward a user by the number of times they provided witnesses.
Now, we reward them for each unit of gas they provided.

* chore: s/Fraud/FraudProof/g